+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.7.4] - 2025-XX-XX
 - Fix issue [#373](https://github.com/intersystems/language-server/issues/373): Add go to definition and hover support for `UrlMap` Call and Forward attribute values
 - Fix issue [#374](https://github.com/intersystems/language-server/issues/374): `undefined` items appearing in completion lists for XML XData
+- Fix issue [#375](https://github.com/intersystems/language-server/issues/375): Retrieve password from Server Manager case-sensitively
 
 ## [2.7.3] - 2025-05-05
 - Fix issue [#366](https://github.com/intersystems/language-server/issues/366): Language Server can return invalid DocumentSymbols in some rare circumstances